# app.js
//. app.js
var express = require( 'express' ),
    bodyParser = require( 'body-parser' ),
    { Configuration, OpenAIApi } = require( 'openai' ),
    app = express();

require( 'dotenv' ).config();

app.use( express.static( __dirname + '/public' ) );
app.use( bodyParser.urlencoded( { extended: true } ) );
app.use( bodyParser.json() );
app.use( express.Router() );

var IGNORE_PHRASE = 10;  //. 結果の最初のフレーズがこの長さ以下だったら無視する

var settings_cors = 'CORS' in process.env ? process.env.CORS : '';  //. "http://localhost:8080,https://xxx.herokuapp.com"
app.all( '/*', function( req, res, next ){
  if( settings_cors ){
    var origin = req.headers.origin;
    if( origin ){
      var cors = settings_cors.split( " " ).join( "" ).split( "," );

      //. cors = [ "*" ] への対応が必要
      if( cors.indexOf( '*' ) > -1 ){
        res.setHeader( 'Access-Control-Allow-Origin', '*' );
        res.setHeader( 'Access-Control-Allow-Methods', '*' );
        res.setHeader( 'Access-Control-Allow-Headers', '*' );
        res.setHeader( 'Vary', 'Origin' );
      }else{
        if( cors.indexOf( origin ) > -1 ){
          res.setHeader( 'Access-Control-Allow-Origin', origin );
          res.setHeader( 'Access-Control-Allow-Methods', '*' );
          res.setHeader( 'Access-Control-Allow-Headers', '*' );
          res.setHeader( 'Vary', 'Origin' );
        }
      }
    }
  }
  next();
});

app.get( '/ping', function( req, res ){
  res.contentType( 'application/json; charset=utf-8' );

  res.write( JSON.stringify( { status: true, message: 'PONG' }, null, 2 ) );
  res.end();
});

var settings_apikey = 'API_KEY' in process.env ? process.env.API_KEY : '';
var settings_organization = 'ORGANIZATION' in process.env ? process.env.ORGANIZATION : '';
var configuration = new Configuration({ apiKey: settings_apikey, organization: settings_organization });
var openai = new OpenAIApi( configuration );
//console.log( openai );

app.get( '/api/engines', async function( req, res ){
  res.contentType( 'application/json; charset=utf-8' );

  var result = await openai.listEngines();
  res.write( JSON.stringify( { status: true, result: result.data.data }, null, 2 ) );
  res.end();
});

app.get( '/api/models', async function( req, res ){
  res.contentType( 'application/json; charset=utf-8' );

  var result = await openai.listModels();
  res.write( JSON.stringify( { status: true, result: result.data.data }, null, 2 ) );
  res.end();
});

app.get( '/api/model/:id', async function( req, res ){
  res.contentType( 'application/json; charset=utf-8' );

  var id = req.params.id;
  var result = await openai.retrieveModel( id );
  res.write( JSON.stringify( { status: true, result: result.data }, null, 2 ) );
  res.end();
});

app.post( '/api/complete', async function( req, res ){
  res.contentType( 'application/json; charset=utf-8' );
  var model = ( req.body.model ? req.body.model : 'text-davinci-003' );
  var max_tokens = ( req.body.max_tokens ? parseInt( req.body.max_tokens ) : 4000 );
  var prompt = req.body.prompt;

  var option = {
    model: model,
    prompt: prompt,
    //messages: [
    //  { role: "user", content: prompt }
    //],
    max_tokens: max_tokens
  };
  if( req.body.temperature ){
    option.temperature = parseFloat( req.body.temperature );
  }
  if( req.body.top_p ){
    option.top_p = parseFloat( req.body.top_p );
  }
  if( req.body.n ){
    option.n = parseInt( req.body.n );
  }

  try{
    //var result = await openai.createCompletion( option );
    var result = await callWithProgress( progressingCompletion, option, 5 ); //. #1
    var answer = result.data.choices[0].text;

    //. 最初の "\n\n" 以降が正しい回答？
    var tmp = answer.split( "\n\n" );
    if( tmp.length > 1 && tmp[0].length < IGNORE_PHRASE ){
      tmp.shift();
      answer = tmp.join( "\n\n" );
    }

    res.write( JSON.stringify( { status: true, result: answer }, null, 2 ) );
    res.end();
  }catch( err ){
    //console.log( {err} );
    //console.log( err.result.response );
    var status_code = ( err.response && err.response.status ? err.response.status : ( err.result && err.result.response && err.result.response.status ? err.result.response.status : 400 ) ); //. #1
    var status_text = ( err.response && err.response.statusText ? err.response.statusText : ( err.result && err.result.response && err.result.response.statusText ? err.result.response.statusText : 'unknown error' ) );
    if( err.result && err.result.response && err.result.response.data && err.result.response.data.error && err.result.response.data.error.message ){
      status_text += '. ' + err.result.response.data.error.message;
    }
    res.status( status_code )
    res.write( JSON.stringify( { status: false, error: status_text }, null, 2 ) );
    res.end();
  }
});

//. #1
const wait = ( ms ) => new Promise( ( res ) => setTimeout( res, ms ) );

const progressingCompletion = async ( option ) => {
	await wait( 10 );
  try{
    var result = await openai.createCompletion( option );
  	return {
      status: true,
		  result: result
  	};
  }catch( e ){
  	return {
      status: false,
		  result: e
  	};
  }
}

const progressingImage = async ( option ) => {
	await wait( 10 );
  try{
    var result = await openai.createImage( option );
  	return {
      status: true,
		  result: result
  	};
  }catch( e ){
  	return {
      status: false,
		  result: e
  	};
  }
}

const progressingChatCompletion = async ( option ) => {
	await wait( 10 );
  try{
    var result = await openai.createChatCompletion( option );
  	return {
      status: true,
		  result: result
  	};
  }catch( e ){
  	return {
      status: false,
		  result: e
  	};
  }
}

const progressingEdit = async ( option ) => {
	await wait( 10 );
  try{
    var result = await openai.createEdit( option );
  	return {
      status: true,
		  result: result
  	};
  }catch( e ){
  	return {
      status: false,
		  result: e
  	};
  }
}

const callWithProgress = async ( fn, option, maxdepth = 7, depth = 0 ) => {
	const result = await fn( option );

	// check completion
	if( result.status ){
		// finished
		return result.result;
	}else{
		if( depth > maxdepth ){
			throw result;
		}
		await wait( Math.pow( 2, depth ) * 10 );
	
		return callWithProgress( fn, option, maxdepth, depth + 1 );
	}
}

app.post( '/api/image', async function( req, res ){
  res.contentType( 'application/json; charset=utf-8' );
  var n = ( req.body.n ? parseInt( req.body.n ) : 1 );
  var size = ( req.body.size ? req.body.size : '256x256' );
  var format = ( req.body.format ? req.body.format : 'url' /* 'url' or 'b64_json' */ );  //. response_format
  var prompt = req.body.prompt;

  var option = {
    prompt: prompt,
    n: n,
    size: size,
    response_format: format
  };

  //var result = await openai.createImage( option );
  var result = await callWithProgress( progressingImage, option, 5 ); //. #1

  res.write( JSON.stringify( { status: true, result: result['data']['data'] }, null, 2 ) );
  res.end();
});

//. #2
app.post( '/api/chat', async function( req, res ){
  res.contentType( 'application/json; charset=utf-8' );
  var role = ( req.body.role ? req.body.role : 'user' );
  var model = ( req.body.model ? req.body.model : 'gpt-3.5-turbo' );
  var max_tokens = ( req.body.max_tokens ? parseInt( req.body.max_tokens ) : 4000 );
  var content = ( req.body.content ? req.body.content : '' );

  var option = {
    model: model,
    messages: [
      { role: role, content: content }
    ],
    max_tokens: max_tokens
  };
  if( req.body.temperature ){
    option.temperature = parseFloat( req.body.temperature );
  }
  if( req.body.top_p ){
    option.top_p = parseFloat( req.body.top_p );
  }
  if( req.body.n ){
    option.n = parseInt( req.body.n );
  }

  try{
    //var result = await openai.createChatCompletion( option );
    var result = await callWithProgress( progressingChatCompletion, option, 5 ); //. #1
    //console.log( {result} );
    var answer = result.data.choices[0].message.content;

    /*
    //. 最初の "\n\n" 以降が正しい回答？
    var tmp = answer.split( "\n\n" );
    if( tmp.length > 1 && tmp[0].length < IGNORE_PHRASE ){
      tmp.shift();
      answer = tmp.join( "\n\n" );
    }
    */

    res.write( JSON.stringify( { status: true, result: answer }, null, 2 ) );
    res.end();
  }catch( err ){
    console.log( {err} );
    //console.log( err.result.response );
    var status_code = ( err.response && err.response.status ? err.response.status : ( err.result && err.result.response && err.result.response.status ? err.result.response.status : 400 ) ); //. #1
    var status_text = ( err.response && err.response.statusText ? err.response.statusText : ( err.result && err.result.response && err.result.response.statusText ? err.result.response.statusText : 'unknown error' ) );
    if( err.result && err.result.response && err.result.response.data && err.result.response.data.error && err.result.response.data.error.message ){
      status_text += '. ' + err.result.response.data.error.message;
    }
    res.status( status_code )
    res.write( JSON.stringify( { status: false, error: status_text }, null, 2 ) );
    res.end();
  }
});

app.post( '/api/edit', async function( req, res ){
  res.contentType( 'application/json; charset=utf-8' );
  var model = ( req.body.model ? req.body.model : 'text-davinci-edit-001' );
  var input = ( req.body.input ? req.body.input : '' );
  var instruction = ( req.body.instruction ? req.body.instruction : '' );

  var option = {
    model: model,
    input: input,
    instruction: instruction
  };
  if( req.body.temperature ){
    option.temperature = parseFloat( req.body.temperature );
  }
  if( req.body.top_p ){
    option.top_p = parseFloat( req.body.top_p );
  }
  if( req.body.n ){
    option.n = parseInt( req.body.n );
  }

  try{
    //var result = await openai.createEdit( option );
    var result = await callWithProgress( progressingEdit, option, 5 ); //. #1
    //console.log( result.data );
    var answer = result.data.choices[0].text;

    //. 最初の "\n\n" 以降が正しい回答？
    var tmp = answer.split( "\n\n" );
    if( tmp.length > 1 && tmp[0].length < IGNORE_PHRASE ){
      tmp.shift();
      answer = tmp.join( "\n\n" );
    }

    res.write( JSON.stringify( { status: true, result: answer }, null, 2 ) );
    res.end();
  }catch( err ){
    //console.log( {err} );
    //console.log( err.result.response );
    var status_code = ( err.response && err.response.status ? err.response.status : ( err.result && err.result.response && err.result.response.status ? err.result.response.status : 400 ) ); //. #1
    var status_text = ( err.response && err.response.statusText ? err.response.statusText : ( err.result && err.result.response && err.result.response.statusText ? err.result.response.statusText : 'unknown error' ) );
    if( err.result && err.result.response && err.result.response.data && err.result.response.data.error && err.result.response.data.error.message ){
      status_text += '. ' + err.result.response.data.error.message;
    }
    res.status( status_code )
    res.write( JSON.stringify( { status: false, error: status_text }, null, 2 ) );
    res.end();
  }
});


var port = process.env.PORT || 8080;
app.listen( port );
console.log( "server starting on " + port + " ..." );

module.exports = app;